requirements.txt
scipy==1.2.1
opencv-python==4.2.0.34
Pillow==5.4.1
numpy==1.16.3
keras==2.2.4
matplotlib==3.0.3